refactor(web/todolist): use jq instead of Nix for TODO generation
although it is nice to be able to generate something like this
using Nix, `fromJSON (readFile ...)` with massive input slows
down the pipeline evaluation and can't be done in parallel to
other build jobs. This CL "outsources" this generation to jq
(especially because we don't gain much from using Nix here),
which results in the benefit that we don't have to wait for
the results of ripgrep+jq+readFile+fromJSON, which instead
just gets put into the resulting pipeline, but does not block
pipeline generation itself.

Change-Id: I5baeecc4ccd8c134675616e73ace5d898c452955