feat(kit): Export TVL stack as a josh workspace
This creates a josh workspace called `ext/kit` which exports all
dependencies needed by external users of the TVL stack.

The intention of this is to have a single clonable "virtual repo"
which provides everything needed to run a system like TVL, but without
the projects that we develop *in* TVL.

Right now this is limited to the basic Nix code (readTree, pipeline
generation) and things required for Gerrit integration.

The workspace maps depot paths into slightly different (simpler)
locations inside the exported view, to avoid a nested, mostly empty
structure.

Note: This is not an "external API" or some such of depot, it is
simply a convenience for exporting some of our tooling.

Change-Id: Ied24aaef117fc2e0c188ec742ae7cd993a5babd3