feat(tvl-kit): Use readTree for reading the kit
This adds more of the logic exposed in the kit to its own build
pipeline, by relying on readTree to read the kit and expose besadii.

Note that the depot nixpkgs pin has been remapped to '.nixpkgs' to
avoid re-exposing it through readTree. I think doing it this way is
cleaner than additional nesting + skip-subtree or some sort of
`builtins.removeAttrs`-based hack.

Change-Id: I0c38688a9652499b34c1ed6e63aa89df005c66aa