feat(nix/buildkite): reflect (direct) deps between drvs in pipeline
Using builtins.getContext we can inspect the derivations a derivation
depends on directly. This allows us to emit dependencies between
buildkite steps if the derivations they represent depend on each other.
The dependencies prevent cases where multiple started jobs would try to
build a certain derivation, so all but one would just spend time waiting
on a lock, decreasing parallelization of the pipeline.

The dependency detection in this CL is quite limited, however. With
builtins.getContext we'll only ever be able to inspect the /direct/
dependencies of the derivation's build environment. If a target in the
pipeline depends on another target indirectly (e. g. via a writeText
build script derivation which isn't exposed as its own target), we fail
to detect the dependency. For this reason, the dependencies for e.g.
buildLisp derivations and NixOS system configurations won't be detected
at the moment.

Change-Id: I11f1fc96010c625b98ca2fddaf6854d33486579e