refactor(3p/nixpkgs): Pin nixpkgs sources via niv
There's no need for us to reinvent the wheel here.

niv pins are in //third_party/sources, and niv commands need to be run
with `niv -s third_party/sources/sources.json` to operate on the
correct file.

Note that niv by default wants to put the loader file in
//nix/sources.nix. This file has been moved to
//third_party/sources/default.nix which makes niv throw a warning, but
everything still works as intended.

Change-Id: I2b2a6f8edf33c429a6d7be9d174ba1996d9a0193