feat(ops/auto-deploy): Prefer precise deploy schedule
Try calling `rebuild-system` on a machine that has `auto-deploy` enabled, and it
will do 2x back-to-back rebuilds. That's because `OnActiveSec=1`, so when
systemd restarts the timers, this triggers an auto-deploy.

So switching from interval scheduling to calendar scheduling fixes this, but it
also offers a more expressive API for controlling releases.

Change-Id: I437b35054d9e62e8ff998a052b24068df760015f