feat(wpcarro/redshift): Support redshift (again) 🎉
So it turns out that there's a NixOS *and* a home-manager module for
redshift. While I couldn't get the NixOS module to WAI, the home-manager version
seems to work just fine.

I'm hard-coding the lat/lon for now, because I still can't get "geoclue2" to
work, but I'm not interested in debugging that at the moment.

Change-Id: I7b44df44dc845ec3ad49e649019df459fc0bca56