refactor(rust-crates-advisory): move report generation into script
This script is somewhat usable by humans (it even has a help screen!)
and can be reused in //users/sterni/nixpkgs-crate-holes. We are using
bash since that allows us to exit with the actual exit code of
cargo-audit - something that's not possible in execline. skarnet told me
at some point that it's won't or should not be implemented or whatever.

Change-Id: I3331ae8222a20e23b8e30dc920ab48af78f0247c