fix(rules.pl): Only allow fast-forwarded subtree merges
When committing subtree merges with an out-of-date canon parent,
Gerrit will synthesize an additional merge commit.

This becomes a problem because the `change-merged` hook which is used
to trigger besadii is invoked on the commit of the CL, which is *not*
the commit that ends up in refs/heads/canon.

This breaks assumptions about creation of revision numbers, which
leads to CI failures of those pipelines and missing revisions.

This didn't actually happen to us, but was discovered in a TVL-like
setup.

Change-Id: I27810042a7515f40a0ad1db42c4830887180ffab
Reviewed-on: https://cl.tvl.fyi/c/depot/+/5231
Reviewed-by: grfn <grfn@gws.fyi>
Reviewed-by: sterni <sternenseemann@systemli.org>
Tested-by: tazjin <tazjin@tvl.su>