feat(depotfmt): Check & format Rust code with rustfmt
The rustfmt configuration of the depot is moved to `rustfmt.toml` (it
is recognised more reliably from this path than from the hidden
.rustfmt.toml).

Nested configuration is possible, but not always used. I've changed
some settings that create hard to read code for my users folder
specifically, we'll see if that is reliably used in CI or not.

If not, we can exclude paths with custom config from the top-level
check and figure out a more granular way of doing it.

Change-Id: I385ce3ef529bda28fac03bfba86fc204c81b8a61