fix(wpcarro/playbooks): update first-of-the-month.org
TL;DR:
- remove stale entries
- add entry for buying TSLA for 401k
- (partially) normalize habits.org by referencing ./first-of-the-month.org

Background:

Earlier today I automated as much of my investments as I can...

- crypto account auto-buys crypto 2x/mo
- Fidelity doesn't support auto-purchases of individual stocks, so I opened a
  Robinhood account to auto-purchase stock 2x/mo

The only remaining thing that AFAIK I *cannot* automate is auto-stock-purchases
for TD Ameritrade (my SDBO 401k account). I don't think I can transfer this to
Robinhood because Vanguard controls which brokerages they're compatible
with. This should still be a big time-saver.

As such, it's time to rely on playbooks for this, so I'm dusting-off the
first-of-the-month.org playbook.

Change-Id: I545f8de20a0a30cac597400c4114b4549671a91d