feat(ops/journaldriver): Support systemd LoadCredential=
With this change, if journaldriver is passed a relative path as the
location of the application credentials, it will attempt to resolve
this path within the systemd `CREDENTIALS_DIRECTORY` (if present).

This supports use-cases with systemd's LoadCredential= functionality.

An alternative way of doing this would be to wrap journaldriver in a
shell script. The limitation here is that systemd can't actually
template this path into the unit file, it's only available at
runtime (which seems like an oversight).

Doing it in here seems cleaner, especially since this program is
already systemd-specific.

Change-Id: I5ce0dd61bfddacbd337fc9dfff58ff16cfa1cce1