chore(tgsa): Remove shortening of image links
The link shortening was in place because posts with image URLs that
were too long would not get recognised as valid image links.

astral has fixed this after my report:

  https://forums.somethingawful.com/showthread.php?threadid=3999866

I verified this with the example post that has a bunch of images and
such in it.

Thanks astral!

Change-Id: I60e90fd52020ee5464c931a4dde91382c8a149ab
Reviewed-on: https://cl.tvl.fyi/c/depot/+/5481
Reviewed-by: tazjin <tazjin@tvl.su>
Autosubmit: tazjin <tazjin@tvl.su>
Tested-by: BuildkiteCI