fix(3p/nixpkgs): prevent pkgs.path being dumped to the store
Paths have for better or worse (really not sure which) no knowledge if
they are in the Nix store or not. This means that they have to be dumped
to store when used in string interpolation like is done in
//ops/nixos.nix. Unfortunately, this caused evaluation of any system
in depot to dump nixpkgs to store, with significant memory and disk
space usage.

We can easily prevent this by overlaying path to be the same as
nixpkgsSrc which Nix can know is already in store.

Before:

nix-repl> users.sterni.machines.edwin.config.nix.nixPath
[ "nixos=/nix/store/hd10pafli40h31x2d40s8ki5azplyykf-fnix5y8gpd262sc7p952dwip15hidfcd-nixpkgs-src" "nixpkgs=/nix/store/hd10pafli40h31x2d40s8ki5azplyykf-fnix5y8gpd262sc7p952dwip15hidfcd-nixpkgs-src" ]

After:

nix-repl> users.sterni.machines.edwin.config.nix.nixPath
[ "nixos=/nix/store/fnix5y8gpd262sc7p952dwip15hidfcd-nixpkgs-src" "nixpkgs=/nix/store/fnix5y8gpd262sc7p952dwip15hidfcd-nixpkgs-src" ]

Change-Id: I9d3292cc44c7976bfb2fcacfb8687baf2e02cebf