fix(tools/magrathea): disallow the empty target for shell
The “empty target” is not something that’s a useful concept for a dev
shell. Where is the difference between that and `//`?

When you run (nix-expr-for-target (empty-target)), you get

```
"with builtins; (import /home/philip/depot {})"'
```

and when you pass that to `nix-shell -E`, you get this unhelpful
error:

```
error: expression does not evaluate to a derivation (or a set or list
of those)
```

Change-Id: I11cecf3cb99eaa7c1f826503b1bac45a472cfd68