feat(rules.pl): Only allow autosubmit for owner or clbot
The autosubmit label can currently lead to submission of changes that
were uploaded by someone other than clbot or the owner of the change,
see b/167.

This could lead to situations where a malicious user could inject code
into a CL that is in the process of being autosubmitted.

With this change, the Autosubmit label is only allowed after checking
the uploader of the patchset.

This was tested through the submit rule testing SSH interface, however
we can't test whether this actually has the desired effect that way.
It's possible that the `may(_)` option only affects the UI state, for
example, but not the actual label on the change (if present).

We will only find out by merging it ...

It should also be noted that the implementation is a bit ugly because
Gerrit's Prolog does not have an append/3 function and I had to pick
between reimplementing that or letting the new predicate do the
appending.

Change-Id: I6bdcd9be19afc1aece5f3eb98b88156477f1f17b