feat(tools/magrathea): Add devshell command
The `devshell` command is a simple helper to create a development
shell with magrathea.

A development shell differs from a normal shell in that it is expected
to contain all tools required to develop the target, wheras the normal
`shell` will open the nix-shell for the project’s derivation, which
usually misses tooling for development (e.g. a compiler with local
documentation like `haskellPackages.ghcWithHoogle`).

If you want a devshell for your project, you have to expose it as
`passthru.devshell` from your project’s derivation, which `mg
devshell` will call `nix-shell` on (so you’d use e.g. `pkgs.mkShell`
to create it).

Change-Id: I7a3835d662f877d266942753aabfc6fcc4e0d3cd